Move breakpoints logic to other class
