+++ src/neo-vm/ExecutionEngine.cs
@@ -1,11 +1,11 @@
-﻿using Neo.VM.Types;
-using System;
+﻿using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
 using System.Runtime.CompilerServices;
 using System.Security.Cryptography;
+using Neo.VM.Types;
 using VMArray = Neo.VM.Types.Array;
 
 namespace Neo.VM
@@ -57,7 +57,6 @@ namespace Neo.VM
         private bool is_stackitem_count_strict = true;
 
         private readonly IScriptTable table;
-        private readonly Dictionary<byte[], HashSet<uint>> break_points = new Dictionary<byte[], HashSet<uint>>(new HashComparer());
 
         public IScriptContainer ScriptContainer { get; }
         public ICrypto Crypto { get; }
@@ -75,16 +74,6 @@ namespace Neo.VM
             this.Crypto = crypto;
             this.table = table;
             this.Service = service;
-        }
-
-        public void AddBreakPoint(byte[] script_hash, uint position)
-        {
-            if (!break_points.TryGetValue(script_hash, out HashSet<uint> hashset))
-            {
-                hashset = new HashSet<uint>();
-                break_points.Add(script_hash, hashset);
-            }
-            hashset.Add(position);
         }
 
         #region Limits
@@ -224,11 +213,6 @@ namespace Neo.VM
                 catch
                 {
                     State = VMState.FAULT;
-                }
-                if (State == VMState.NONE && InvocationStack.Count > 0)
-                {
-                    if (break_points.Count > 0 && break_points.TryGetValue(CurrentContext.ScriptHash, out HashSet<uint> hashset) && hashset.Contains((uint)CurrentContext.InstructionPointer))
-                        State = VMState.BREAK;
                 }
             }
         }
@@ -1305,17 +1289,6 @@ namespace Neo.VM
 
         protected virtual bool PreExecuteInstruction()
         {
-            return true;
-        }
-
-        public bool RemoveBreakPoint(byte[] script_hash, uint position)
-        {
-            if (!break_points.TryGetValue(script_hash, out HashSet<uint> hashset))
-                return false;
-            if (!hashset.Remove(position))
-                return false;
-            if (hashset.Count == 0)
-                break_points.Remove(script_hash);
             return true;
         }
